fix(avx2): restore scalar bit-exactness for VIF, PSNR-HVS, and ADM (ADR-0138) (#1228)
Three AVX2 paths diverged from their scalar references due to
float-precision narrowing at accumulation boundaries:

Bug 1 — vif_statistic_s_avx2: `sigma_max_inv` used `255.0f * 255.0f`
(float) as denominator; scalar reference uses `255.0 * 255.0` (double).
Division at float precision produces 1–5e-6 divergence from frame 3 of
src01_hrc00/hrc01. Fix: cast result to float after double-precision divide.

Bug 2 — calc_psnrhvs_avx2: `compute_masks()` called `sqrt(s_mask *
s_gvar)` (float × float) without the `(double)` cast that the scalar
(psnr_hvs.c:351) uses, causing ~3.77e-7 divergence on the Cb channel at
frame 0. Fix: `sqrt((double)b->s_mask * b->s_gvar)`.

Bug 3 — adm_cm_avx2: h-accumulator conversion had a spurious inner float
cast `(float)((float)accum_h / pow(...))` vs the scalar's `(float)(accum_h
/ pow(...))`. Premature int64→float narrowing (accum_h ≈ 8.4e14) loses
mantissa bits before the double-precision divide, producing up to 498M ULP
error (max 2.3e-6 in adm_scale3) on src01 frame 3. Fix: remove inner cast.

Add three new SIMD bit-exactness tests (ADR-0245 pattern):
  test_vif_simd      — scalar vs AVX2 stat, rel-tol 1e-6, 5 cases
  test_psnr_hvs_simd — inline scalar vs calc_psnrhvs_avx2, rel-tol 1e-12
  test_integer_adm_simd — arithmetic regression + adm_cm_avx2 smoke

All three test executables registered in libvmaf/test/meson.build under
suite ['fast', 'simd'], x86/x86_64 only.

+++ libvmaf/src/feature/x86/adm_avx2.c
@@ -2730,7 +2730,16 @@ float adm_cm_avx2(AdmBuffer *buf, int w, int h, int src_stride, int csf_a_stride
      * hence pending is 57-shift's done based on width and height
      */
 
-    float f_accum_h = (float)((float)accum_h / pow(2, (52 - shift_xhcub - shift_inner_accum)));
+    /* ADR-0138 bit-exactness fix: scalar (integer_adm.c:2009) computes
+     *   f_accum_h = (float)(accum_h / pow(2, ...))
+     * which is int64 → double → divide → float.  The AVX2 path originally
+     * had a spurious inner `(float)` cast on accum_h only:
+     *   (float)((float)accum_h / pow(...))
+     * This converts int64 to float first — losing mantissa bits for large
+     * accum_h — before the division, producing up to 498M ULP error on
+     * adm_scale3 (max 2.3e-6 in the final score).  Mirror the scalar
+     * exactly for all three accumulators. */
+    float f_accum_h = (float)(accum_h / pow(2, (52 - shift_xhcub - shift_inner_accum)));
     float f_accum_v = (float)(accum_v / pow(2, (52 - shift_xvcub - shift_inner_accum)));
     float f_accum_d = (float)(accum_d / pow(2, (57 - shift_xdcub - shift_inner_accum)));
