+++ libvmaf/src/feature/x86/psnr_hvs_avx2.c
@@ -404,13 +404,16 @@ static void compute_masks(psnr_hvs_block *b, const float mask[8][8])
             b->d_mask += b->dct_d[i * 8 + j] * b->dct_d[i * 8 + j] * mask[i][j];
         }
     }
-    /* ADR-0141: `sqrt` (double) matches the scalar reference's float->double
-     * promotion before sqrt; switching to `sqrtf` would diverge from the
-     * bit-exact contract. */
-    // NOLINTNEXTLINE(performance-type-promotion-in-math-fn)
-    b->s_mask = sqrt(b->s_mask * b->s_gvar) / 32.f;
-    // NOLINTNEXTLINE(performance-type-promotion-in-math-fn) ADR-0141 as above.
-    b->d_mask = sqrt(b->d_mask * b->d_gvar) / 32.f;
+    /* ADR-0138 bit-exactness fix: scalar reference at third_party/xiph/psnr_hvs.c:351
+     * does `sqrt((double)s_mask * s_gvar) / 32.f` — the explicit cast makes the
+     * multiplication double-precision before sqrt.  Without the cast, `s_mask *
+     * s_gvar` is a float-precision multiply (both operands are float), and the
+     * result differs by ~3.77e-7 on the Cb channel at frame 0.
+     * Preserve the scalar's (double) cast to maintain byte-for-byte parity. */
+    // NOLINTNEXTLINE(performance-type-promotion-in-math-fn) ADR-0138: matches scalar.
+    b->s_mask = (float)(sqrt((double)b->s_mask * b->s_gvar) / 32.0);
+    // NOLINTNEXTLINE(performance-type-promotion-in-math-fn) ADR-0138: matches scalar.
+    b->d_mask = (float)(sqrt((double)b->d_mask * b->d_gvar) / 32.0);
     if (b->d_mask > b->s_mask) {
         b->s_mask = b->d_mask;
     }
